requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
openai
pytesseract
pdf2image
python-docx
openpyxl
python-pptx
moviepy
Pillow
chromadb
tiktoken
pydantic
python-dotenv
ffmpeg-python
pytube
openai-whisper
accelerate
torch
vllm
transformers
serpapi
scrapfly-sdk
dotenv
google-search-results
pytz
brotli
pymupdf
frontend
tools
langchain_chroma
bitsandbytes